refactor: [Node] обновление классов до версии 2.0
