+++ common/src/main/java/com/github/jon7even/telegram/constants/DefaultSystemMessagesToSend.java
@@ -12,4 +12,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class DefaultSystemMessagesToSend {
     public static final String WE_NOT_SUPPORT = "мы еще не поддерживаем";
+    public static final String ERROR_TO_EXECUTION_FOR_USER = "Произошла ошибка в процессе ответа пользователю Error:";
 }
