+++ node/src/main/java/com/github/jon7even/NodeApp.java
@@ -7,11 +7,13 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
  * Основной класс запуска сервиса "Узел", управляется фреймворком SpringBoot
  *
  * @author Jon7even
- * @version 1.0
+ * @version 2.0
  */
 @SpringBootApplication
 public class NodeApp {
+
     public static void main(String[] args) {
         SpringApplication.run(NodeApp.class);
     }
+
 }
