+++ node/src/main/java/com/github/jon7even/cache/UserDataCache.java
@@ -4,13 +4,14 @@ import com.github.jon7even.dto.company.CompanyBuildingDto;
 import com.github.jon7even.telegram.BotState;
 
 /**
- * Интерфейс сервиса отвечающего за кеш пользователя (конструирование новых моделей, хранение состояния бота для
- * конкретного пользователя)
+ * Интерфейс сервиса отвечающего за кэш пользователя
  *
  * @author Jon7even
- * @version 1.0
+ * @version 2.0
+ * @apiNote конструирование новых моделей, хранение состояния бота для конкретного пользователя
  */
 public interface UserDataCache {
+
     void setBotStateForCacheUser(Long userId, BotState botState);
 
     void setCompanyForCacheUser(Long userId, CompanyBuildingDto companyBuildingDto);
