+++ node/src/main/java/com/github/jon7even/configuration/RabbitConfiguration.java
@@ -9,10 +9,11 @@ import org.springframework.context.annotation.Configuration;
  * Конфигурация для работы с брокером очередей RabbitMq
  *
  * @author Jon7even
- * @version 1.0
+ * @version 2.0
  */
 @Configuration
 public class RabbitConfiguration {
+
     @Bean
     public MessageConverter jsonMessageConverter() {
         return new Jackson2JsonMessageConverter();
