--- node/src/main/java/com/github/jon7even/dto/competitor/CompetitorShortDto.java
@@ -1,20 +1,0 @@
-package com.github.jon7even.dto.competitor;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * Класс DTO для краткого представления конкурента
- *
- * @author Jon7even
- * @version 1.0
- */
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class CompetitorShortDto {
-    private String name;
-}
