--- node/src/main/java/com/github/jon7even/entity/company/EmployeeEntity.java
@@ -1,38 +1,0 @@
-package com.github.jon7even.entity.company;
-
-import com.github.jon7even.entity.user.UserEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-/**
- * Класс описывающий сущность представителя компании
- *
- * @author Jon7even
- * @version 1.0
- */
-@Entity
-@Builder
-@Setter
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "employees", schema = "public")
-public class EmployeeEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @Column(name = "full_name", nullable = false)
-    private String name;
-
-    @Column(name = "added_on", nullable = false)
-    private LocalDateTime registeredOn;
-
-    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.MERGE)
-    @JoinColumn(name = "creator_id", nullable = false)
-    @ToString.Exclude
-    private UserEntity creator;
-}
