--- node/src/main/java/com/github/jon7even/entity/competitor/CompetitorEntity.java
@@ -1,38 +1,0 @@
-package com.github.jon7even.entity.competitor;
-
-import com.github.jon7even.entity.user.UserEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-/**
- * Класс описывающий сущность конкурента
- *
- * @author Jon7even
- * @version 1.0
- */
-@Entity
-@Builder
-@Setter
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "competitors", schema = "public")
-public class CompetitorEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @Column(name = "name_competitor", nullable = false)
-    private String name;
-
-    @Column(name = "created_on", nullable = false)
-    private LocalDateTime created;
-
-    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.MERGE)
-    @JoinColumn(name = "creator_id", nullable = false)
-    @ToString.Exclude
-    private UserEntity creator;
-}
