--- node/src/main/java/com/github/jon7even/mapper/CompanyMapper.java
@@ -1,36 +1,0 @@
-package com.github.jon7even.mapper;
-
-import com.github.jon7even.dto.company.CompanyBuildingDto;
-import com.github.jon7even.entity.company.CompanyEntity;
-import com.github.jon7even.entity.gift.GiftEntity;
-import com.github.jon7even.entity.user.UserEntity;
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-import org.mapstruct.factory.Mappers;
-
-import java.time.LocalDateTime;
-
-/**
- * Интерфейс для маппинга DTO и сущностей компаний
- *
- * @author Jon7even
- * @version 1.0
- */
-@Mapper(componentModel = "spring")
-public interface CompanyMapper {
-    CompanyMapper INSTANCE = Mappers.getMapper(CompanyMapper.class);
-
-    @Mapping(target = "id", ignore = true)
-    @Mapping(source = "companyBuildingDto.name", target = "name")
-    @Mapping(source = "companyBuildingDto.totalSum", target = "totalSum")
-    @Mapping(source = "gift", target = "gift")
-    @Mapping(source = "now", target = "created")
-    @Mapping(source = "userEntity", target = "creator")
-    @Mapping(source = "companyBuildingDto.isGiven", target = "isGiven")
-    @Mapping(target = "updated", ignore = true)
-    @Mapping(target = "given", ignore = true)
-    CompanyEntity toEntityFromBuildingDto(CompanyBuildingDto companyBuildingDto,
-                                          UserEntity userEntity,
-                                          GiftEntity gift,
-                                          LocalDateTime now);
-}
