+++ node/src/main/java/com/github/jon7even/service/in/message/ReplyMessageService.java
@@ -4,10 +4,11 @@ package com.github.jon7even.service.in.message;
  * Интерфейс сервиса конвертации сообщений
  *
  * @author Jon7even
- * @version 1.0
+ * @version 2.0
  * @apiNote Использует таг сообщения из файла messages_ru_RU.properties и преобразует его в ответ
  */
 public interface ReplyMessageService {
+
     String getReplyText(String replyText);
 
     String getReplyText(String replyText, Object... args);
