+++ node/src/main/java/com/github/jon7even/service/in/message/impl/LocaleMessageServiceImpl.java
@@ -1,5 +1,6 @@
-package com.github.jon7even.service.in.message;
+package com.github.jon7even.service.in.message.impl;
 
+import com.github.jon7even.service.in.message.LocaleMessageService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Service;
@@ -7,14 +8,16 @@ import org.springframework.stereotype.Service;
 import java.util.Locale;
 
 /**
- * Реализация сервиса языковой локализации
+ * Реализация сервиса языковой локализации {@link LocaleMessageService}
  *
  * @author Jon7even
- * @version 1.0
+ * @version 2.0
  */
 @Service
 public class LocaleMessageServiceImpl implements LocaleMessageService {
+
     private final Locale locale;
+
     private final MessageSource messageSource;
 
     public LocaleMessageServiceImpl(@Value("${localeTag}") String localeTag, MessageSource messageSource) {
