+++ node/src/main/java/com/github/jon7even/service/in/message/impl/ReplyMessageServiceImpl.java
@@ -1,17 +1,20 @@
-package com.github.jon7even.service.in.message;
+package com.github.jon7even.service.in.message.impl;
 
+import com.github.jon7even.service.in.message.LocaleMessageService;
+import com.github.jon7even.service.in.message.ReplyMessageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 /**
- * Реализация сервиса конвертации сообщений
+ * Реализация сервиса конвертации сообщений {@link ReplyMessageService}
  *
  * @author Jon7even
- * @version 1.0
+ * @version 2.0
  */
 @Service
 @RequiredArgsConstructor
 public class ReplyMessageServiceImpl implements ReplyMessageService {
+
     private final LocaleMessageService localeMessageService;
 
     @Override
