+++ node/src/main/java/com/github/jon7even/telegram/constants/DefaultMessageError.java
@@ -7,10 +7,11 @@ import lombok.NoArgsConstructor;
  * Утилитарный класс для типовых служебных сообщений, которые отправляются пользователю при ошибках сервиса
  *
  * @author Jon7even
- * @version 1.0
+ * @version 2.0
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class DefaultMessageError {
+
     public static final String ERROR_TO_SEND = "Извините, произошла ошибка: ";
 
     public static final String ERROR_SEND_TEXT = "что-то произошло с сервисом ответов на сообщения, "
